Finally compiling

+++ src/multi/facets/EdgeFacet.sol
@@ -12,7 +12,7 @@ contract EdgeFacet {
     function setEdge(
         address token0,
         address token1,
-        uint256 amplitude,
+        uint128 amplitude,
         int24 lowTick,
         int24 highTick
     ) external {
